Prevent `parseTarget` in resolver conn wrapper from clobbering target
The last split on the target's endpoint overrides the authority as well
as the endpoint with an empty string if the endpoint doesn't contain a
slash (this is how `split2` above works). In our case the target is an
SRV record following the format "_service-ssl._protocol.hostname.tld".
After `parseTarget` is done, the resolver target is all empty.
The crux is that `parseTarget` is careful _if_ the target cannot be
split on "://" but that won't work in our case because we're using a
custom resolver and that needs the scheme to pick the correct one.